+++ src/all/background_page/error/passboltApiFetchError.js
@@ -9,6 +9,7 @@ class PassboltApiFetchError extends Error {
   constructor(message, data) {
     super(message);
     this.data = data || {};
+    this.name = 'PassboltApiFetchError';
   }
 }