+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - PASSBOLT-3312: As GM adding a user to a group I should see a relevant feedback in case of network/proxy errors
 - PASSBOLT-3316: As LU Sharing a password I should see a loading feedback when the extension is requesting the API
 - PASSBOLT-3318: As LU I should retrieve a secret when I'm copying it
+- PASSBOLT-3319: As LU I should retrieve a secret when I'm editing it
 
 ## [2.4.6] - 2018-12-18
 ### Fix